python-amd-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
transformers==4.35.0
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
numpy==1.24.3
python-multipart==0.0.6
pydub==0.25.1
python-dotenv==1.0.0
soundfile==0.12.1
httpx==0.25.2
requests==2.31.0
py7zr==0.20.6

